refactor: upvotes and downvotes merged into votes
The previous implementation of `Post` contained two `Set`s, one
for the upvoters and one for the downvoters. However, while
adding an upvote I did not check if the post had already been
downvoted, or viceversa: with `votes` I can do that concurrently.

Moreover, I finally understood that `putIfAbsent` returns the
previous value of the entry